Remove bugs - Navbar
