+++ frontend/src/appComponents/Herosection.tsx
@@ -1,15 +1,6 @@
 "use client";
 
-import { useState } from "react";
-import { Dock, DockIcon } from "@/components/magicui/dock";
-import cross from "../assets/cross-icon.svg";
-import burger from "../assets/burger-menu.svg";
-
 const Hero: React.FC = () => {
-  const [isMenuOpen, setIsMenuOpen] = useState(false);
-  const toggleMenu = () => {
-    setIsMenuOpen(!isMenuOpen);
-  };
 
   return (
     <div className="relative z-10 flex flex-col w-full overflow-hidden text-white bg-background md:shadow-xl">
@@ -31,72 +22,6 @@ const Hero: React.FC = () => {
           </button>
         </div>
       </div>
-
-      
-      <div className="z-20 transform transition-all duration-500 fixed top-8 hover:left-[34%] left-[35%] w-[550px] hover:w-[600px]">
-        <Dock
-          direction="middle"
-          magnification={60}
-          distance={100}
-          className="hidden lg:flex items-center w-full text-base border rounded-lg shadow-xl bg-white/10 backdrop-blur-sm border-white/20"
-        >
-          <DockIcon>
-            <p className="flex items-center font-semibold transition-all duration-300 transform size-full">
-              Home
-            </p>
-          </DockIcon>
-          <DockIcon>
-            <p className="flex items-center font-semibold transition-all duration-300 transform size-full">
-              Clubs
-            </p>
-          </DockIcon>
-          <DockIcon>
-            <p className="flex items-center font-semibold transition-all duration-300 transform size-full">
-              Events
-            </p>
-          </DockIcon>
-          <DockIcon>
-            <p className="flex items-center font-semibold transition-all duration-300 transform size-full">
-              Hiring
-            </p>
-          </DockIcon>
-        </Dock>
-      </div>
-
-      <div className="z-20 transform transition-all duration-500 fixed top-4 right-4 md:hidden">
-        <button onClick={toggleMenu}>
-          <img
-            src={isMenuOpen ? cross : burger}
-            width={50}
-            height={50}
-            alt={isMenuOpen ? "Close Menu" : "Open Menu"}
-          />
-        </button>
-      </div>
-
-     
-     {isMenuOpen && (
-        <div className="fixed inset-0 bg-white flex flex-col items-center justify-center text-6xl font-bold font-pixeboy z-[1001]">
-          <button className="absolute top-5 right-5" onClick={toggleMenu}>
-            <img
-              src={cross}
-              width={75}
-              height={75}
-              alt="Close Menu"
-            />
-          </button>
-        
-          <ul className="flex flex-col gap-8 mt-12 text-black">
-            <li className="hover:underline cursor-pointer text-center ">Home</li>
-            <li className="hover:underline cursor-pointer text-center ">Clubs</li>
-            <li className="hover:underline cursor-pointer text-center ">Events</li>
-            <li className="hover:underline cursor-pointer text-center ">Hiring</li>
-          </ul>
-        </div>
-      )}
-
-
-
     </div>
   );
 };
